DailyBuzz-Django/DailyBuzz/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.2
django-heroku==0.3.1
djangorestframework==3.16.0
frozenlist==1.5.0
geojson==2.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.4.3
newsapi-python==0.2.7
packaging==24.2
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.1.0
python-weather==2.0.7
requests==2.32.3
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
yarl==1.19.0
